requirements.txt
requests
python-decouple
bs4
lxml
pyTelegramBotApi
